moved mailer to helper folder as per request
