+++ helpers/mailer.js
@@ -1,5 +1,5 @@
 const nodemailer = require('nodemailer');
-const config = require('./config');
+const config = require('../config');
 
 const transporter = nodemailer.createTransport(config.email);
