+++ routes/pnid.js
@@ -12,7 +12,7 @@ const passport = require('passport');
 const userMiddleware = require('../middleware/authentication');
 const apiHelper = require('../helpers/api');
 const utilHelper = require('../helpers/util');
-const mailer = require('../mailer');
+const mailer = require('../helpers/mailer');
 const config = require('../config.json');
 const Recaptcha = require('express-recaptcha').Recaptcha;
 const recaptcha = new Recaptcha(config.recaptcha.siteKey, config.recaptcha.secretKey);
